pyproject.toml
[project]
name = "rpg-soundboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyqt6>=6.10.2",
    "python-vlc>=3.0.21203",
]
